Depend on itertools:
We intend to use merge() to combine multiple values within a filter element
(all the different authors, but now combined and sorted by Record ID)

We then intend to do an intersection across multiple filter elements

In this way, our indexes can all be (TYPE+ID) where ID starts with timestamp
so these are in sorted order (but going forward in time). We just flip our
cmp operators so make it sort backwards in time.  And we need no compound
indexes, saving a lot of space.